backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio]
motor
pydantic
openai
python-dotenv
